requirements.txt
absl-py==2.2.2
accelerate==0.26.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
angelcommunity==2.0.0
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
bimlpa==0.1.2
bitsandbytes==0.44.1
bleach==6.1.0
blis==0.4.1
boto3==1.35.78
botocore==1.35.78
catalogue==1.0.2
cdlib==0.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
comm==0.2.3
contourpy==1.1.1
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.10
datasets==3.1.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
demon==2.0.6
Deprecated==1.2.15
dill==0.3.8
docker-pycreds==0.4.0
docstring-parser==0.15
duckdb==1.2.0
dynetx==0.3.2
eva-lcd==0.1.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
faiss-gpu==1.7.2
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gensim==3.8.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
igraph==0.11.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonlines==1.2.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
Levenshtein==0.25.1
lion-pytorch==0.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
matrix-client==0.4.0
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
narwhals==1.42.1
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
openai==0.27.8
opencv-python==4.11.0.86
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandasai==3.0.0b11
pandocfilters==1.5.1
parso==0.8.4
peft==0.13.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==4.3.6
plotly==6.2.0
pooch==1.8.2
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
PuLP==3.1.1
pure_eval==0.2.3
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyGithub==2.5.0
Pygments==2.17.2
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-igraph==0.11.8
python-json-logger==3.3.0
python-Levenshtein==0.25.1
python-louvain==0.16
python_paris==0.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==27.0.1
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rouge-score==0.1.2
rpds-py==0.20.1
s3transfer==0.10.4
sacremoses==0.1.1
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
shtab==1.6.5
six==1.17.0
sklearn==0.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
sqlglot==25.34.1
srsly==1.0.7
stack-data==0.6.3
sympy==1.13.3
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.18.1
texttable==1.7.0
thinc==7.3.1
threadpoolctl==3.5.0
thresholdclustering==1.1
tinycss2==1.2.1
tokenizers==0.20.3
tomli==2.2.1
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.0.0
trl==0.7.4
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tyro==0.6.0
tzdata==2025.1
UNKNOWN==0.45.1.dev0
uri-template==1.3.0
urllib3==2.2.3
wandb==0.19.1
wasabi==0.10.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.0
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
